Revert "feat: simple staking sunset modal (#11022)"
This reverts commit 7fcbd5e29101db668472786737a8e9b0d81ca09e.
