+++ apps/web/src/components/Modal/AffiliateSunsetModal.tsx
@@ -1,11 +1,23 @@
 import { useTranslation } from '@pancakeswap/localization'
 import { LinkExternal, Text, useModal } from '@pancakeswap/uikit'
-import { Trans } from 'components/CustomTrans'
-import { useCallback, useEffect } from 'react'
+import { ReactNode, useCallback, useEffect } from 'react'
 
 import DisclaimerModal from 'components/DisclaimerModal'
 import { useUserIsInAffiliateListData } from 'hooks/useAffiliateSunsetList'
 import { useUserAcknowledgement } from 'hooks/useUserAcknowledgement'
+
+const transRegex = /(%[^%]+%)/
+
+function Trans({ text, data = {} }: { text: string; data?: { [key: string]: ReactNode } }) {
+  const parts = text.split(transRegex)
+  return parts.map((p) => {
+    if (!transRegex.test(p)) {
+      return p
+    }
+    const key = p.replace(/%/g, '')
+    return data[key] || p
+  })
+}
 
 export function AffiliateSunsetModal() {
   const { t } = useTranslation()
