+++ apps/web/src/pages/_app.tsx
@@ -5,7 +5,6 @@ import GlobalCheckClaimStatus from 'components/GlobalCheckClaimStatus'
 import { PageMeta } from 'components/Layout/Page'
 import { AffiliateExpiredModal } from 'components/Modal/AffiliateExpiredModal'
 import { AffiliateSunsetModal } from 'components/Modal/AffiliateSunsetModal'
-import { SimpleStakingSunsetModal } from 'components/Modal/SimpleStakingSunsetModal'
 import { NetworkModal } from 'components/NetworkModal'
 import { FixedSubgraphHealthIndicator } from 'components/SubgraphHealthIndicator/FixedSubgraphHealthIndicator'
 import TransactionsDetailModal from 'components/TransactionDetailModal'
@@ -199,7 +198,6 @@ const App = ({ Component, pageProps }: AppPropsWithLayout) => {
       <ZKSyncAirdropModalWithAutoPopup />
       <AffiliateExpiredModal />
       <AffiliateSunsetModal />
-      <SimpleStakingSunsetModal />
       <VercelToolbar />
     </ProductionErrorBoundary>
   )
