+++ packages/localization/src/config/translations.json
@@ -3668,13 +3668,5 @@
   "Springboard": "Springboard",
   "PancakeSwap MEV Guard": "PancakeSwap MEV Guard",
   "Create & Launch Your Token in minutes, on our ": "Create & Launch Your Token in minutes, on our ",
-  "SpringBoard!!": "SpringBoard!!",
-  "Action Required: Simple Staking Product Retirement": "Action Required: Simple Staking Product Retirement",
-  "Our %page% Product will retire on %time%, at 00:00 UTC. Please ensure you withdraw all your funds before this date.": "Our %page% Product will retire on %time%, at 00:00 UTC. Please ensure you withdraw all your funds before this date.",
-  "This is part of our ongoing efforts to streamline offerings and focus on more sustainable, impactful products for our community.": "This is part of our ongoing efforts to streamline offerings and focus on more sustainable, impactful products for our community.",
-  "To withdraw your funds:": "To withdraw your funds:",
-  "Visit the %page%": "Visit the %page%",
-  "Simple Staking Dashboard": "Simple Staking Dashboard",
-  "Select your stake and withdraw": "Select your stake and withdraw",
-  "Learn more here": "Learn more here"
+  "SpringBoard!!": "SpringBoard!!"
 }
